move logger to separate module
